+++ S4-DemoApplication/src/main.tsx
@@ -3,6 +3,9 @@ import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App'
 import { AuthProvider } from './contexts/AuthContext'
+import { initSentry } from './lib/sentry'
+
+initSentry()
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>